fix(mock-draft): port all recent server changes into party/draft-room.ts (the deployed file)
Every prior party-side change landed in party/mock-draft.ts, but
partykit.json points main at party/draft-room.ts — mock-draft.ts was
dead code. That's why undo, my-rank, and the instant auto-pick
behaviour never showed up in the deployed server even after pnpm
party:deploy succeeded.

Ports everything into the real file:
- Session-room init handler now persists rankedLists, rankingAssignments
  and defaultRankingSource (previously only the legacy flat ranked-
  players list was stored — that's why per-team rank assignments and
  My Rank were silently ignored).
- autoPick() consults the assigned source first, then defaultSource,
  then the legacy list, then any source with stock left. Old sessions
  with only rankedPlayerIds still work via the legacy fallback.
- scheduleNextPick() replaces startTimer at all five transition call
  sites (join-starts-draft, handleStart, handleResume, handleReset,
  post-makePick). Rule: creator's slot → full user timer; any other
  slot → instant auto-pick via Promise microtask so the current pick-
  made broadcast flushes first.
- handleUndo message type + handler: pops the last pick back to an
  empty slot, decrements currentPickIndex, reopens status completed →
  active if undo lands on the final pick, schedules the next pick.

Also deletes party/mock-draft.ts outright so we can't accidentally
edit the wrong file again.
